src/sensor_software/executable/requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.3
Adafruit-SHT31==1.0.2
blessings==1.7
python-dotenv==0.9.1
six==1.11.0
websocket-client==0.53.0
